+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## NEXT
+- add `Downloader::headers`
+
 ## 0.2.0 - 2025-11-21
 - add `DownloaderBuilder::delay`
 - update dependencies